requirements.txt
numpy==1.18.5
matplotlib==3.3.0
pillow==7.2.0
selenium==3.141.0
requests==2.24.0
onnxruntime

uvicorn
fastapi
fastapi_restful
python-multipart

